Refactoriza la gestión de notificaciones en `NotificationsApiService` y `NotificationsMonitorFacade`, implementando un sistema de caché para mejorar el rendimiento. Se añaden métodos para marcar notificaciones como leídas y se optimizan las consultas en `NotificationsQueryDomain`. Además, se actualizan los componentes de la interfaz para incluir acciones de eliminación y se mejora la experiencia del usuario al gestionar notificaciones.

+++ supabase/functions/delete-user/index.ts
@@ -22,6 +22,10 @@ async function deleteUser(email: string) {
     throw { message: 'Usuario no encontrado en tabla Accounts', status: 404 };
   }
 
+  if (account.IsOwner) {
+    throw { message: 'El usuario no puede ser eliminado', status: 400 };
+  }
+
   // 2. Borrar roles en AccountRoles
   const { error: rolesError } = await supabase.from(SupabaseTables.AccountRoles).delete().eq('AccountId', account.id);
